+++ CHANGELOG.md
@@ -67,6 +67,7 @@ Developers may need to change their code
 
 ### Corrected Busg
 
+* Fixed: [Error when changing choice/case with different structure](https://github.com/clicon/clixon/issues/568)
 * Fixed: [Clixon handle if-feature incorrectly](https://github.com/clicon/clixon/issues/555)
 * Fixed: [Clixon fails to load yang with extension](https://github.com/clicon/clixon/issues/554)
 * Fixed: Double top-levels in xmldb_get that could occur with xpath containing choice.